Create "ecdsattack" package and put test in dedicated files
