+++ ecdsattack/attacks/f.py
@@ -12,10 +12,8 @@ s_bad * (r_good * x + digest) / s_good = r_bad * x + digest
 """
 
 from ecdsa.curves import NIST256p
-from ecdsa.util import randrange
 
-from .common import _ecdsa_sign
-from signature import Signature
+from ..common import Signature
 
 
 def F(good: Signature, bad: Signature) -> int:
@@ -27,26 +25,3 @@ def F(good: Signature, bad: Signature) -> int:
     if denom % n == 0:
         return 0
     return (digest * (bad.s - good.s) * pow(denom, -1, n)) % n
-
-
-def _ecdsa_sign_with_fault(msg: int, x: int, k: int):
-    n = NIST256p.order
-
-    # Usually, Q = k * NIST256p.generator and r = Q.x() % n. Here, fault in r
-    r = randrange(n)
-    kinv = pow(k, n - 2, n)
-
-    s = (kinv * (msg + x * r)) % n
-    return Signature(msg, r, s)
-
-
-def test_f():
-    n = NIST256p.order
-
-    msg = randrange(n)
-    x = randrange(n)
-    k = randrange(n)
-    sig1 = _ecdsa_sign(msg, x, k)
-    sig2 = _ecdsa_sign_with_fault(msg, x, k)
-
-    assert F(sig1, sig2) == x
