+++ ecdsattack/attacks/fc2.py
@@ -19,10 +19,8 @@ s_good2*r_bad2(4)-s_good1*r_bad1(5) => 0 = s_good2 * r_bad2 * s_bad1 * r_good1 *
 """
 
 from ecdsa.curves import NIST256p
-from ecdsa.util import randrange
 
-from .common import _ecdsa_sign
-from signature import Signature
+from ..common import Signature
 
 
 def FC2(good1: Signature, bad1: Signature, good2: Signature, bad2: Signature) -> int:
@@ -38,20 +36,3 @@ def FC2(good1: Signature, bad1: Signature, good2: Signature, bad2: Signature) ->
     if denom % n == 0:
         return 0
     return (num * pow(denom, -1, n)) % n
-
-
-def test_fc2():
-    n = NIST256p.order
-
-    msg1 = randrange(n)
-    msg2 = randrange(n)
-    k1 = randrange(n)
-    k2 = randrange(n)
-    x = randrange(n)
-    x2 = randrange(n)  # fault on x
-
-    sig1 = _ecdsa_sign(msg1, x, k1)
-    sig2 = _ecdsa_sign(msg1, x2, k1)
-    sig3 = _ecdsa_sign(msg2, x, k2)
-    sig4 = _ecdsa_sign(msg2, x2, k2)
-    assert FC2(sig1, sig2, sig3, sig4) == x
