+++ ecdsattack/attacks/fc3.py
@@ -24,10 +24,8 @@ s_bad2*(3)-s_good2*(4) => (similarly) s_good2 * e = x * (s_bad2 * r_good2 - s_go
 """
 
 from ecdsa.curves import NIST256p
-from ecdsa.util import randrange
 
-from .common import _ecdsa_sign
-from signature import Signature
+from ..common import Signature
 
 
 def FC3(
@@ -50,31 +48,3 @@ def FC3(
     if denom % n == 0:
         return 0
     return (num * pow(denom, -1, n)) % n
-
-
-def _ecdsa_sign_with_fault(msg: int, x: int, k: int, e: int) -> Signature:
-    n = NIST256p.order
-
-    Q = k * NIST256p.generator
-    r = Q.x() % n
-    kinv = pow(k, n - 2, n)
-    s = (kinv * (0 + x * r)) % n
-    return Signature(msg, r, s)
-
-
-def test_fc3():
-    n = NIST256p.order
-
-    msg1 = randrange(n)
-    msg2 = randrange(n)
-    k1 = randrange(n)
-    k2 = randrange(n)
-    x = randrange(n)
-
-    e = randrange(n)
-
-    sig1 = _ecdsa_sign(msg1, x, k1)
-    sig2 = _ecdsa_sign_with_fault(msg1, x, k1, e)
-    sig3 = _ecdsa_sign(msg2, x, k2)
-    sig4 = _ecdsa_sign_with_fault(msg2, x, k2, e)
-    assert FC3(sig1, sig2, sig3, sig4) == x
