+++ ecdsattack/attacks/fc4.py
@@ -19,10 +19,8 @@ s_good2*(5)-s_good1*(6) => 0 = s_good2 * s_bad1 * r_good1 * x + s_good2 * s_bad1
 """
 
 from ecdsa.curves import NIST256p
-from ecdsa.util import randrange
 
-from .common import _ecdsa_sign
-from signature import Signature
+from ..common import Signature
 
 
 def FC4(good1: Signature, bad1: Signature, good2: Signature, bad2: Signature) -> int:
@@ -36,31 +34,3 @@ def FC4(good1: Signature, bad1: Signature, good2: Signature, bad2: Signature) ->
     if denom % n == 0:
         return 0
     return (num * pow(denom, -1, n)) % n
-
-
-def _ecdsa_sign_with_fault(msg: int, k: int, e: int) -> Signature:
-    n = NIST256p.order
-
-    Q = k * NIST256p.generator
-    r = Q.x() % n
-    kinv = pow(k, n - 2, n)
-    s = (kinv * e) % n
-    return Signature(msg, r, s)
-
-
-def test_fc4():
-    n = NIST256p.order
-
-    msg1 = randrange(n)
-    msg2 = randrange(n)
-    k1 = randrange(n)
-    k2 = randrange(n)
-    x = randrange(n)
-
-    e = randrange(n)
-
-    sig1 = _ecdsa_sign(msg1, x, k1)
-    sig2 = _ecdsa_sign_with_fault(msg1, k1, e)
-    sig3 = _ecdsa_sign(msg2, x, k2)
-    sig4 = _ecdsa_sign_with_fault(msg2, k2, e)
-    assert FC4(sig1, sig2, sig3, sig4) == x
