+++ ecdsattack/common.py
