+++ pwn_fault.py
@@ -10,21 +10,12 @@ import random
 import shutil
 import string
 import subprocess
-from typing import List, Optional
+from typing import Optional
 
 from ecdsa.curves import NIST256p
 from ecdsa.ellipticcurve import Point
 
-from attacks.f import F
-from attacks.fc1 import FC1
-from attacks.fc2 import FC2
-from attacks.fc3 import FC3
-from attacks.fc4 import FC4
-from attacks.fc5 import FC5
-from attacks.fdc1 import FDC1
-from attacks.fdc2 import FDC2
-from attacks.fdc3 import FDC3
-from signature import Signature
+from ecdsattack import recover_key, Signature
 
 ORIGINAL_FILENAME_A = "main_a"
 ORIGINAL_FILENAME_B = "main_b"
@@ -32,30 +23,6 @@ ECDSA_SIG_SIZE = 256 // 8 * 2
 
 DIGEST_A = 0xAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
 DIGEST_B = 0xBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB
-
-
-def recover_key(
-    correct_sigs: List[Signature], faulty_sigs: List[Signature]
-) -> List[int]:
-    res = []
-
-    # first try using F, only needing a couple:
-    for i in range(len(correct_sigs)):
-        res.append(F(correct_sigs[i], faulty_sigs[i]))
-
-    if len(correct_sigs) == 2 and len(faulty_sigs) == 2:
-        c0, c1 = correct_sigs
-        f0, f1 = faulty_sigs
-
-        res.append(FC1(c0, f0, c1, f1))
-        res.append(FC2(c0, f0, c1, f1))
-        res.append(FC3(c0, f0, c1, f1))
-        res.append(FC4(c0, f0, c1, f1))
-        res.append(FC5(f0, f1))
-        res.append(FDC1(c0, f0, c1, f1))
-        res.append(FDC2(c0, f0, c1, f1))
-        res.append(FDC3(c0, f0, c1, f1))
-    return res
 
 
 def inject_and_run(origin_file_name: str, fault=None):
@@ -133,6 +100,7 @@ def ecdsa_fault_attack(challenge_id: int, fast_mode=False):
         compile_challenge(ORIGINAL_FILENAME_B, challenge_id)
         correct_sigs.append(get_signature(ORIGINAL_FILENAME_B, DIGEST_B))
         file_size = os.path.getsize(ORIGINAL_FILENAME_B)
+    print("Got original signatures")
 
     nb_no_effect = 0
     nb_crashes = 0
@@ -164,19 +132,15 @@ def ecdsa_fault_attack(challenge_id: int, fast_mode=False):
             nb_crashes += 1
         else:
             print("Found fault:", faulty_sigs)
-            print("Trying to recover the key...")
-            dd = recover_key(correct_sigs, faulty_sigs)
-
-            for d in dd:
-                print("Trying", d)
-                if d * NIST256p.generator == public_key:
-                    print("Found correct public point:", public_key)
-                    print("Found private key:", d)
-                    print("In hex:", hex(d))
-                    print(f"Fault: index={hex(byte_index)}, value={hex(byte_value)}")
-                    print("# crashes = ", nb_crashes)
-                    print("# faults without effect = ", nb_no_effect)
-                    return
+            d = recover_key(NIST256p.generator, public_key, correct_sigs, faulty_sigs)
+            if d:
+                print("Found correct public point:", public_key)
+                print("Found private key:", d)
+                print("In hex:", hex(d))
+                print(f"Fault: index={hex(byte_index)}, value={hex(byte_value)}")
+                print("# crashes = ", nb_crashes)
+                print("# faults without effect = ", nb_no_effect)
+                return
 
     print("No exploitable fault found")
     print("# crashes = ", nb_crashes)
